Added a bunch of features

+++ src/main/java/purge.java
@@ -27,38 +27,37 @@ import java.util.stream.Collectors;
 public class purge {
 
     public static void purge(MessageReceivedEvent event){
-        IMessage[] messagesToDelete;
-        MessageList msgDel;
         List<IMessage> msgDel2 = new ArrayList<IMessage>();
         int toDel = 0;
         String[] msg = event.getMessage().toString().split("\\s");
         if(msg[1].matches("\\d+")) {
-            toDel = Integer.parseInt(msg[1]);
-            if(toDel > 100){
-                toDel = 100;
-            }
-            msgDel2 = event.getMessage().getChannel().getMessages().stream().sequential().sorted((a, b) -> b.getCreationDate().compareTo(a.getCreationDate())).limit(toDel).collect(Collectors.toList());
+            if(whitelist.checkWh(event)) {
+                toDel = Integer.parseInt(msg[1]);
+                if (toDel > 100) {
+                    toDel = 100;
+                }
+                msgDel2 = event.getMessage().getChannel().getMessages().stream().sequential().sorted((a, b) -> b.getCreationDate().compareTo(a.getCreationDate())).limit(toDel).collect(Collectors.toList());
             /*
             * Loop IChannel.getMessages().get(index)
             * */
-            try {
-                event.getMessage().getChannel().getMessages().bulkDelete(msgDel2);
-            } catch (DiscordException e) {
-                e.printStackTrace();
-            } catch (RateLimitException e) {
-                e.printStackTrace();
-            } catch (MissingPermissionsException e) {
-                e.printStackTrace();
+
+                try {
+                    event.getMessage().getChannel().getMessages().bulkDelete(msgDel2);
+                } catch (DiscordException |RateLimitException | MissingPermissionsException e) {
+                    e.printStackTrace();
+                }
+            }else{
+                try {
+                    event.getMessage().getChannel().sendMessage("You are not whitelisted!");
+                } catch (DiscordException |RateLimitException | MissingPermissionsException e) {
+                    e.printStackTrace();
+                }
             }
 
         } else {
             try {
                 event.getMessage().getChannel().sendMessage("Incorrect Parameter! Parameter have to consists of only Numbers!");
-            } catch (MissingPermissionsException e) {
-                e.printStackTrace();
-            } catch (RateLimitException e) {
-                e.printStackTrace();
-            } catch (DiscordException e) {
+            } catch (DiscordException |RateLimitException | MissingPermissionsException e) {
                 e.printStackTrace();
             }
         }
